actualizacion de componentes

+++ src/components/patients/PatientsList.jsx
@@ -1,10 +1,11 @@
 /* eslint-disable no-unused-vars */
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
+import { UserContext } from '../Helpers/userContext';
 import { Table } from "antd";
 import Header from '../Header';
 import Sidebar from '../Sidebar';
 import { onShowSizeChange, itemRender } from '../Pagination';
-import { blogimg2, imagesend, pdficon, pdficon3, pdficon4, refreshicon, searchnormal, plusicon } from '../imagepath';
+import {  imagesend, pdficon, pdficon3, pdficon4, refreshicon, searchnormal, plusicon } from '../imagepath';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
 import Swal from 'sweetalert2';
@@ -12,31 +13,15 @@ import FeatherIcon from "feather-icons-react/build/FeatherIcon";
 
 
 const PatientsList = () => {
-  const [selectedRowKeys, setSelectedRowKeys] = useState([]);
+  const {usuarioLogged} = useContext(UserContext);
   const [pacientes, setPacientes] = useState([]);
-  const [pacientesFiltrados, setPacientesFiltrados] = useState([]); // Para la búsqueda
+  const [pacientesFiltrados, setPacientesFiltrados] = useState([]);
   const [searchQuery, setSearchQuery] = useState("");
-  const [selectedPatientId, setSelectedPatientId] = useState(null);
-
-  const user = JSON.parse(localStorage.getItem('user'));
-
-  const onSelectChange = (newSelectedRowKeys) => {
-    console.log("selectedRowKeys changed: ", newSelectedRowKeys);
-    setSelectedRowKeys(newSelectedRowKeys);
-  };
-
-  const API_URL = import.meta.env.VITE_REACT_APP_API_URL;
-  
-  const rowSelection = {
-    selectedRowKeys,
-    onChange: onSelectChange,
-  };
 
   const getPacientes = async () => {
     try {
       const response = await axios.get(`${import.meta.env.VITE_REACT_APP_API_URL}/pacientes`);
-      // console.log(response.data);
-      setPacientes(response.data);
+      setPacientes(response.data.reverse());
     } catch (error) {
       console.error("Error al cargar los datos: ", error);
     }
@@ -56,39 +41,49 @@ const PatientsList = () => {
 
   // Función para manejar la eliminación
   const handleDelete = async (id) => {
-    try {
-      await axios.delete(`${import.meta.env.VITE_REACT_APP_API_URL}/pacientes/${id}`);
-
-      // Si la eliminación fue exitosa
-      const updatedPacientes = pacientes.filter(paciente => paciente.id_paciente !== id);
-      setPacientes(updatedPacientes);
-      setPacientesFiltrados(updatedPacientes);
-      setSelectedPatientId(null);  // Limpiar el estado
-
-      // Cerrar el modal y mostrar alerta de éxito
-      Swal.fire({
-        icon: 'success',
-        title: '¡Paciente eliminado!',
-        text: 'El paciente ha sido eliminado con éxito.',
-        showConfirmButton: false,
-        timer: 1500
-      }).then(() => {
-        // Aquí cerramos el modal
-        const closeButton = document.querySelector('[data-bs-dismiss="modal"]');
-        if (closeButton) closeButton.click(); // Cierra el modal
-      });
-    } catch (error) {
-      // En caso de error, mostramos un mensaje
-      console.error('Error al eliminar el paciente:', error);
-      Swal.fire({
-        icon: 'error',
-        title: '¡Error!',
-        text: 'Hubo un problema al eliminar el paciente.',
-        confirmButtonText: 'Cerrar'
-      });
-    }
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'No podrás revertir esta acción.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.delete(`${import.meta.env.VITE_REACT_APP_API_URL}/pacientes/${id}`);
+  
+          // Si la eliminación fue exitosa
+          const updatedPacientes = pacientes.filter(paciente => paciente.id_paciente !== id);
+          setPacientes(updatedPacientes);
+          setPacientesFiltrados(updatedPacientes);
+  
+          Swal.fire({
+            icon: 'success',
+            title: '¡Paciente eliminado!',
+            text: 'El paciente ha sido eliminado con éxito.',
+            showConfirmButton: false,
+            timer: 1500,
+          }).then(() => {
+            // Aquí cerramos el modal
+            const closeButton = document.querySelector('[data-bs-dismiss="modal"]');
+            if (closeButton) closeButton.click(); // Cierra el modal
+          });
+        } catch (error) {
+          // En caso de error, mostramos un mensaje
+          console.error('Error al eliminar el paciente:', error);
+          Swal.fire({
+            icon: 'error',
+            title: '¡Error!',
+            text: 'Hubo un problema al eliminar el paciente.',
+            confirmButtonText: 'Cerrar',
+          });
+        }
+      }
+    });
   };
-
 
   // Refrescar los datos
   const handleRefresh = () => {
@@ -102,7 +97,6 @@ const PatientsList = () => {
     {
       title: "Nombre Completo",
       dataIndex: "nombre",
-      sorter: (a, b) => a.nombre.length - b.nombre.length,
       render: (text, record) => (
         <h2 className="profile-image">
           {record.nombre_completo}
@@ -110,9 +104,8 @@ const PatientsList = () => {
       ),
     },
     {
-      title: "# identidad",
+      title: "# Identidad",
       dataIndex: "num_doc",
-      sorter: (a, b) => a.num_doc.length - b.num_doc.length,
       render: (text, record) => (
         <h2 className="profile-image">
           {record.numero_identidad}
@@ -122,7 +115,6 @@ const PatientsList = () => {
     {
       title: "Edad",
       dataIndex: "edad",
-      sorter: (a, b) => a.edad.length - b.edad.length,
       render: (text, record) => (
         <h2 className="profile-image">
           {record.edad}
@@ -132,7 +124,6 @@ const PatientsList = () => {
     {
       title: "Teléfono",
       dataIndex: "telefono",
-      sorter: (a, b) => a.telefono.length - b.telefono.length,
       render: (text, record) => (
         <h2 className="profile-image">
           {record.telefono}
@@ -141,8 +132,7 @@ const PatientsList = () => {
     },
     {
       title: "Correo",
-      dataIndex: "corre",
-      sorter: (a, b) => a.correo.length - b.correo.length,
+      dataIndex: "correo",
       render: (text, record) => (
         <h2 className="profile-image">
           {record.correo_electronico}
@@ -150,11 +140,12 @@ const PatientsList = () => {
       ),
     },
     {
-      title: "",
+      title: "Acciones",
       dataIndex: "FIELD8",
       render: (text, record) => (
         <div className="text-end">
           <div className="dropdown dropdown-action">
+           
             <Link
               to="#"
               className="action-icon dropdown-toggle"
@@ -172,20 +163,14 @@ const PatientsList = () => {
                 <i className="far fa-edit me-2" />
                 Editar
               </Link>
-              {/* Solo podra eliminar si el usuario loggueado es administrador. */}
-              {user.id_rol == 3 ? (
-                // Botón de Eliminar
-                <Link
+              {usuarioLogged?.rol?.permisos.some((permiso) => permiso.nombre === "eliminar") && (
+                <button
                   className="dropdown-item"
-                  to="#"
-                  data-bs-toggle="modal"
-                  data-bs-target="#delete_patient"
-                  onClick={() => setSelectedPatientId(record.id_paciente)}
+                  onClick={() => handleDelete(record.id_paciente)}
                 >
                   <i className="fa fa-trash-alt m-r-5"></i> Eliminar
-                </Link>
-
-              ) : null}
+                </button>
+              )}
             </div>
           </div>
         </div>
@@ -276,28 +261,6 @@ const PatientsList = () => {
                     />
                   </div>
                 </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div id="delete_patient" className="modal fade delete-modal" role="dialog">
-        <div className="modal-dialog modal-dialog-centered">
-          <div className="modal-content">
-            <div className="modal-body text-center">
-              <img src={imagesend} alt="#" width={50} height={46} />
-              <h3>¿Estás seguro de eliminarlo?</h3>
-              <div className="m-t-20">
-                <Link to="#" className="btn btn-white me-2" data-bs-dismiss="modal">
-                  Cerrar
-                </Link>
-                <button
-                  type="button"
-                  className="btn btn-danger"
-                  onClick={() => handleDelete(selectedPatientId)}
-                >
-                  Eliminar
-                </button>
               </div>
             </div>
           </div>
